Moved files out and started adding symbols
